requirements.txt
Django==3.2.13
django-environ==0.8.1
django-cors-headers==3.12.0
drf-yasg==1.20.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dj-rest-auth==2.2.4
django-filter==21.1
django-nested-admin==3.4.0
pyotp==2.6.0
Pillow==9.1.1
pre-commit==2.19.0
gunicorn==20.1.0
django-celery-results==2.3.1
django-redis==5.2.0
flower==1.0.0
django-celery-beat==2.2.1
psycopg2-binary==2.9.3
django-ckeditor==6.4.2
channels==3.0.4
channels-redis==3.4.0
django-debug-toolbar==3.6.0
segno==1.5.2
fcm-django==1.0.12
XlsxWriter==3.0.3
